Move `r_home_setup()` to harp
